+++ Scenes/Common/unknown_signal_button.gd
@@ -1,0 +1,7 @@
+extends Button
+class_name UnknownSignalButton
+
+var sig: int = 0
+
+func _pressed():
+	DictEditMenu.open(sig)