+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 * Deleting dictionary entries now prompts confirmation. Courtesy of [Eearslya](https://github.com/A5TR0spud/dscr_asterclient/commits?author=Eearslya) on Discord. Hell yeah
+* Added buttons to quickly define an unknown frequency when it's used in a transmission
 
 ## 0.4.2